refact: organize server endpoints into `server/api` mod
These modules should make the code base more legibile.

These changes include:
- Moving `serve` current documents endpoint into `server/api/serve`
  module.

  This way we'll centralize all future endpoint logic in the `api` module

- Centralize routing into `routes` mod.

 We'll expect to add new 'static' routing, while also keeping our
 dynamic logic in the same place.

- Centralize global app state into `state` mod.

- Small updates following the re-org of current codebase

refact: rename structs in `state.rs` module to fix linting issues

Address clippy lint warnings/pedantic errors related to generics in function arguments
