+++ src/server/api/mod.rs
@@ -1,2 +1,5 @@
 //! This module contains the API endpoints for the server.
+pub mod routes;
+pub mod serve;
+pub mod state;
 pub mod versions;
